+++ app/src/release/java/com/example/bikini_android/ui/settings/SettingsFragment.kt
@@ -1,0 +1,23 @@
+/*
+ * SettingsFragment.kt 2020. 11. 1
+ *
+ * Copyright 2020 BasicBug. All rights Reserved.
+ *
+ */
+
+package com.example.bikini_android.ui.settings
+
+import android.os.Bundle
+import android.view.View
+import com.example.bikini_android.ui.base.BaseSettingsFragment
+
+/**
+ * @author MyeongKi
+ */
+
+class SettingsFragment : BaseSettingsFragment() {
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        //여기에 필요한 xml 선언 및 준비
+    }
+}
