+++ app/src/release/res/values/arrays.xml
