+++ app/src/release/res/xml/preference_settings_main.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ preference_settings_main.xml 2020. 11. 5
+  ~
+  ~ Copyright 2020 BasicBug. All rights Reserved.
+  -->
+
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+
+</PreferenceScreen>
